restructured the app
