+++ my-app/src/types/types.tsx
@@ -1,18 +1,18 @@
 import React from "react";
 
 export type Questions = {
-   id: string;
-   question: string;
-   points: number;
-   answer: Answer[];
+  id: string;
+  question: string;
+  points: number;
+  options: Answer[];
 };
 
 export type Answer = {
-   option: string;
-   isRight: boolean;
+  option: string;
+  isRight: boolean;
 };
 
 export type Quiz = {
-   quizName: string;
-   questions: Questions[];
+  quizName: string;
+  questions: Questions[];
 };
