+++ my-app/src/utilities/utilities.tsx
@@ -1,0 +1,3 @@
+export function utilities() {
+  return <div>asdasd</div>;
+}
